Test indentation on new line in scripts.
