+++ idea/testData/indentationOnNewline/ScriptAfterExpression.kts
@@ -1,0 +1,4 @@
+package a.b.c
+
+var x = 1
+x = x + 1<caret>
