+++ idea/testData/indentationOnNewline/ScriptAfterExpression_after.kts
@@ -1,0 +1,5 @@
+package a.b.c
+
+var x = 1
+x = x + 1
+<caret>
