+++ idea/testData/indentationOnNewline/ScriptAfterFun.kts
@@ -1,0 +1,5 @@
+package a.b.c
+
+fun fn() {
+
+}<caret>
