+++ idea/testData/indentationOnNewline/ScriptAfterFun_after.kts
@@ -1,0 +1,6 @@
+package a.b.c
+
+fun fn() {
+
+}
+<caret>
