+++ idea/testData/indentationOnNewline/ScriptAfterImport.kts
@@ -1,0 +1,3 @@
+package a.b.c
+
+import x.y.z<caret>
