+++ idea/testData/indentationOnNewline/ScriptAfterImport_after.kts
@@ -1,0 +1,4 @@
+package a.b.c
+
+import x.y.z
+<caret>
