+++ idea/testData/indentationOnNewline/ScriptInsideFun.kts
@@ -1,0 +1,6 @@
+package a.b.c
+
+var x = 2
+fun fn() {
+    x = x + 1<caret>
+}
