+++ idea/testData/indentationOnNewline/ScriptInsideFun_after.kts
@@ -1,0 +1,7 @@
+package a.b.c
+
+var x = 2
+fun fn() {
+    x = x + 1
+    <caret>
+}
