+++ idea/tests/org/jetbrains/jet/formatter/JetTypingIndentationTest.java
@@ -26,8 +26,7 @@ import org.jetbrains.jet.testing.SettingsConfigurator;
 
 import java.io.File;
 
-public class JetTypingIndentationTest extends LightCodeInsightTestCase {
-
+public class JetTypingIndentationTest extends JetTypingIndentationTestBase {
     public void testAfterCatch() {
         doFileNewlineTest();
     }
@@ -87,52 +86,5 @@ public class JetTypingIndentationTest extends LightCodeInsightTestCase {
     public void testWhile() {
         doFileNewlineTest();
     }
-
-    public void doFileNewlineTest() {
-        doNewlineTest(getBeforeFileName(), getAfterFileName());
-    }
-
-    public String getBeforeFileName() {
-        return getTestName(false) + ".kt";
-    }
-
-    public String getAfterFileName() {
-        return getTestName(false) + "_after.kt";
-    }
-
-    public String getInvertedAfterFileName() {
-        return getTestName(false) + "_after_inv.kt";
-    }
-
-    public void doFileSettingNewLineTest() throws Exception {
-        String originalFileText = FileUtil.loadFile(new File(getTestDataPath(), getBeforeFileName()), true);
-
-        SettingsConfigurator configurator = JetFormatSettingsUtil.createConfigurator(originalFileText);
-
-        configurator.configureSettings();
-        doNewlineTest(getBeforeFileName(), getAfterFileName());
-
-        configurator.configureInvertedSettings();
-        doNewlineTest(getBeforeFileName(), getInvertedAfterFileName());
-
-        getSettings().clearCodeStyleSettings();
-    }
-
-    private void doNewlineTest(String beforeFileName, String afterFileName) {
-        configureByFile(beforeFileName);
-        type('\n');
-        checkResultByFile(afterFileName);
-    }
-
-    public static CodeStyleSettings getSettings() {
-        return CodeStyleSettingsManager.getSettings(getProject());
-    }
-
-    @NotNull
-    @Override
-    protected String getTestDataPath() {
-        String testRelativeDir = "indentationOnNewline";
-        return new File(PluginTestCaseBase.getTestDataPathBase(), testRelativeDir).getPath() +
-               File.separator;
-    }
 }
+
